+++ CHANGELOG.md
@@ -26,7 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Security in case of vulnerabilities.
 
 ## [Unreleased]
-- Nil.
+- Improve Worldwide::Config.configure_i18n performance [#301](https://github.com/Shopify/worldwide/pull/301)
 
 ---